requirements.txt
mysql_connector_repackaged==0.3.1
pika==1.3.1
SQLAlchemy==2.0.12
